+++ function-images/tests/stacking-training/driver/main.py
@@ -25,10 +25,12 @@ from __future__ import print_function
 import sys
 import os
 
-# adding python tracing sources to the system path
+# adding python tracing and storage sources to the system path
 sys.path.insert(0, os.getcwd() + '/../proto/')
 sys.path.insert(0, os.getcwd() + '/../../../../utils/tracing/python')
+sys.path.insert(0, os.getcwd() + '/../../../../utils/storage/python')
 import tracing
+import storage
 import helloworld_pb2_grpc
 import helloworld_pb2
 import stacking_pb2_grpc
@@ -196,32 +198,6 @@ class GreeterServicer(helloworld_pb2_grpc.GreeterServicer):
                 log.fatal("Empty XDT config")
             self.XDTconfig = XDTconfig
 
-    def put(self, obj, key):
-        msg = "Driver uploading object with key '" + key + "' to " + self.transferType
-        log.info(msg)
-        with tracing.Span(msg):
-            pickled = pickle.dumps(obj)
-            if self.transferType == S3:
-                s3object = self.s3_client.Object(bucket_name=self.benchName, key=key)
-                s3object.put(Body=pickled)
-            elif self.transferType == XDT:
-                log.fatal("XDT is not supported")
-
-        return key
-
-    def get(self, key):
-        msg = "Driver gets key '" + key + "' from " + self.transferType
-        log.info(msg)
-        with tracing.Span(msg):
-            response = None
-            if self.transferType == S3:
-                obj = self.s3_client.Object(bucket_name=self.benchName, key=key)
-                response = obj.get()
-            elif self.transferType == XDT:
-                log.fatal("XDT is not yet supported")
-
-        return pickle.loads(response['Body'].read())
-
     def train(self, arg: dict) -> dict:
         log.info(f"Invoke Trainer {arg['trainer_id']}")
         channel = grpc.insecure_channel(args.tAddr)
@@ -298,14 +274,14 @@ class GreeterServicer(helloworld_pb2_grpc.GreeterServicer):
     def get_final(self, outputs: dict):
         log.info("Get the final outputs")
 
-        _ = self.get(outputs['model_full_key'])
-        _ = self.get(outputs['meta_predictions_key'])
+        _ = storage.get(outputs['model_full_key'])
+        _ = storage.get(outputs['meta_predictions_key'])
 
     # Driver code below
     def SayHello(self, request, context):
         log.info("Driver received a request")
-
-        dataset_key = self.put(self.dataset, "dataset")
+        
+        dataset_key = storage.put("dataset", self.dataset)
 
         training_responses = self.train_all(dataset_key)
 
@@ -321,6 +297,7 @@ class GreeterServicer(helloworld_pb2_grpc.GreeterServicer):
 def serve():
     transferType = os.getenv('TRANSFER_TYPE', S3)
     if transferType == S3:
+        storage.init("S3", 'vhive-stacking')
         log.info("Using inline or s3 transfers")
         max_workers = int(os.getenv("MAX_SERVER_THREADS", 10))
         server = grpc.server(futures.ThreadPoolExecutor(max_workers=max_workers))